\name{sim.Dvine}
\alias{sim.Dvine}
\title{Flexible Pair-Copula Estimation in D-vines with Penalized
  Splines}
\description{Simulating a p-dimensional vector of the fitted
  p-dimensional D-vine.}
\usage{sim.Dvine(DV,N=NULL)}
\arguments{
  \item{DV}{object of class 'penDvine', result of 'Dvine'.}
  \item{N}{Number of simulated p-dimensional pairs from the fitted
    D-vine 'DV'.}
}

\value{The returing values are simulated values of the current Dvine.}

\author{Christian Schellhase <cschellhase@wiwi.uni-bielefeld.de>}
\references{Flexible Pair-Copula Estimation in D-vines using Bivariate Penalized
  Splines, Kauermann G. and Schellhase C. (2014+), Statistics and Computing (to appear).}
